// service/src/models/create-order-list.model.ts
export interface ICreateOrderList {
	soId: string;
	locationSource: string;
	locationReceipt: string;
	grandTotalAmt?: string;
	userId: string;
	saleCode?: string;
	queueNo: string;
	cusNameOrder: string;
	taxCardId?: string;
	cusMobileNoOrder?: string;
	customerAddress?: string;
	addrNo?: string;
	moo?: string;
	mooban?: string;
	buildingName?: string;
	floor?: string;
	room?: string;
	soi?: string;
	streetName?: string;
	tumbon?: string;
	amphur?: string;
	province?: string;
	zipCode?: string;
	country?: string;
	paymentRemark?: string;
	paymentMethod?: string;
	bankCode?: string;
	installmentTerm?: string;
	installmentRate?: string;
	focCode?: string;
	mobileAisFlg?: string;
	bankAbbr?: string;
	reqMinimumBalance?: string;
	preBookingNo?: string;
	depositAmt?: string;
	qrTransId?: string;
	qrAmt?: string;
	qrAirtimeTransId?: string;
	qrAirtimeAmt?: string;
	qrSecretKey?: string;
	convertToNetwotkType?: string;
	soChannelType: string;
	soDocumentType: string;
	shipCusName?: string;
	shipCusAddr?: string;
	shipCustAddr2?: string;
	storeName?: string;
	shipLocation?: string;
	shipMobileNo?: string;
	clearingType?: string;
	qrOrderId?: string;
	creditCardNo?: string;
	cardExpireDate?: string;
	remarkReceipt?: string;
	receiptType?: string;
	customerType?: string;
	taxBranchNo?: string;
	taxNo?: string;
	remark?: string;
	taxPassportId?: string;
	saleEmpId?: string;
	orderNo?: string;
	kycFlg?: string;
	pointMobileNo?: string;
	pointUssd?: string;
	pointReturnCode?: string;
	pointDiscAmt?: string;
	pointAmt?: string;
	refundSessionId?: string;
	trackingNo?: string;
	voucherNo?: string;
	voucherAmt?: string;
	paymentSubType?: string;
	forwarderName?: string;
	whPickingDate?: string;
	sellerName?: string;
	matCodeCarePlus?: string;
	qrOrderIdCarePlus?: string;
	qrTransIdCarePlus?: string;
	qrAmtCarePlus?: string;
	paymentMethodCarePlus?: string;
	bankAbbrCarePlus?: string;
	bankCodeCarePlus?: string;
	qrSecretKeyCarePlus?: string;
	nextAction?: string;
	packageName?: string;
	productList: IProductListCreateOrder;
}

export interface IProductListCreateOrder {
	productType: string;
	soCompany: string;
	productSubType: string;
	brand: string;
	model: string;
	color?: string;
	matCode?: string;
	priceIncAmt?: string;
	priceDiscountAmt?: string;
	qty: string;
	matAirTime?: string;
	tradeNo?: string;
	ussdCode?: string;
	returnCode?: string;
	cashBackFlg?: string;
	tradeAirtimeId?: string;
	tradeDiscountId?: string;
	orderCode?: string;
	contractTempId?: string;
	serialNo?: string;
	simMobileNo?: string;
	valRate?: string;
	packageType?: string;
	simNetworkType?: string;
	fbbNewFlg?: string;
	listMatFreeGoods: IListMatFreeGoods[];
}

export interface IListMatFreeGoods {
	matCodeFG?: string;
	qtyFG?: string;
	tradeFreeGoodsId?: string;
}